requirements.txt
# AWS Bedrock Integration Dependencies
boto3>=1.34.0
botocore>=1.34.0
python-dotenv>=1.0.0

# Async support
asyncio-throttle>=1.0.2

# Logging and utilities
structlog>=23.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Type hints
typing-extensions>=4.7.0

# Agent orchestration
crewai>=0.40.0
pydantic>=1.10

# Document processing and chunking
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# API server
Flask>=3.0
gunicorn>=21.2